Feat 2.8 (Days off request skeleton, repo and CRUD) and split DoctorUI into submenus
